+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
  * renamed `puli package --add` to `puli package --install`
  * added `--dev` option to `puli package --install` command
+ * added `%dev%` variable to `--format` option of `puli package` command
 
 * 1.0.0-beta6 (2015-08-12)